requirements.txt
celery==4.0.2
cryptography==1.7.2
Flask==0.12
Flask_Cors==3.0.2
Flask_SQLAlchemy==2.1
numpy==1.12.1
SQLAlchemy==1.1.5
pytest-pep8